requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.4.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
dotenv==0.9.9
fastapi==0.116.2
frozenlist==1.8.0
gidgethub==5.4.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
loguru==0.7.3
multidict==6.7.0
packaging==25.0
pluggy==1.5.0
propcache==0.4.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pytest==8.3.5
python-dotenv==1.2.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
uvicorn==0.36.0
yarl==1.22.0
pytest-asyncio